Renamed pgr_contract_blob to contracted_rt.h
