+++ src/contraction/src/contractGraph.c
@@ -39,7 +39,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/debug_macro.h"
 #include "c_common/e_report.h"
 #include "c_common/time_msg.h"
-#include "c_types/pgr_contracted_blob.h"
+#include "c_types/contracted_rt.h"
 #include "c_common/edges_input.h"
 #include "c_common/arrays_input.h"
 #include "./contractGraph_driver.h"
